move models::common to types.rs (#101)
