+++ geoplegma/src/adapters/dggal/grids.rs
@@ -13,7 +13,7 @@ use crate::api::{DggrsApi, DggrsApiConfig};
 use crate::constants::whole_earth_bbox;
 use crate::error::DggrsError;
 use crate::error::dggal::DggalError;
-use crate::models::common::{DggrsName, DggrsUid, RefinementLevel, RelativeDepth, ZoneId, Zones};
+use crate::types::{DggrsName, DggrsUid, RefinementLevel, RelativeDepth, ZoneId, Zones};
 use dggal::DGGRS;
 use dggal_rust::dggal;
 use geo::{Point, Rect};
