+++ geoplegma/src/adapters/h3o/common.rs
@@ -10,7 +10,7 @@
 use crate::{
     api::DggrsApiConfig,
     error::{DggrsError, h3o::H3oError},
-    models::common::{RefinementLevel, Zone, ZoneId, Zones},
+    types::{RefinementLevel, Zone, ZoneId, Zones},
 };
 use geo::{Coord, CoordsIter, GeodesicArea, LineString, Point, Polygon};
 use h3o::{Boundary, CellIndex, LatLng, Resolution};
