+++ geoplegma/src/error/port.rs
@@ -11,7 +11,7 @@ use crate::error::FactoryError;
 use crate::error::dggal::DggalError;
 use crate::error::dggrid::DggridError;
 use crate::error::h3o::H3oError;
-use crate::models::common::{RefinementLevel, RelativeDepth};
+use crate::types::{RefinementLevel, RelativeDepth};
 use std::num::ParseFloatError;
 use thiserror::Error;
